fix: data type in the profits and balance table from float to double in amount

+++ database/migrations/2022_04_01_151122_create_balances_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->engine="InnoDB";
             $table->id();
             $table->bigInteger('user_id')->unsigned();
-            $table->float('amount')->default(0);
+            $table->double('amount')->default(0);
             $table->date('date')->nullable();
             $table->boolean('is_valid')->default(0);
             $table->string('type');
